# server/docs/swagger.yaml
definitions:
...
  api.ProbeCertificateRequest:
    properties:
      domain:
        type: string
    required:
    - domain
    type: object
...
  model.CertCertificate:
    properties:
      ID:
        description: 主键ID
        type: integer
      brand:
        type: string
      category:
        type: string
      certId:
        type: integer
      certStatus:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      domain:
        type: string
      domainExpireAt:
        type: string
      domainStartAt:
        type: string
      domainStatus:
        type: integer
      errorLog:
        type: string
      expireAt:
        type: string
      parentId:
        type: integer
      startAt:
        type: string
      tlsVersion:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.8
paths:
...
  /certCertificate/createCertCertificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 证书模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CertCertificate'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建证书
      tags:
      - CertCertificate
  /certCertificate/deleteCertCertificate:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: query
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除证书
      tags:
      - CertCertificate
  /certCertificate/deleteCertCertificateByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        description: IDs
        in: query
        items:
          type: string
        name: IDs
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除证书
      tags:
      - CertCertificate
  /certCertificate/findCertCertificate:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: query
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.CertCertificate'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询证书
      tags:
      - CertCertificate
  /certCertificate/getCertCertificateList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - in: query
        name: brand
        type: string
      - in: query
        name: category
        type: string
      - in: query
        name: certId
        type: integer
      - in: query
        name: certStatus
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: domain
        type: string
      - in: query
        name: domainExpireAt
        type: string
      - in: query
        name: domainStartAt
        type: string
      - in: query
        name: domainStatus
        type: integer
      - in: query
        name: errorLog
        type: string
      - in: query
        name: expireAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: startAt
        type: string
      - in: query
        name: tlsVersion
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取证书列表
      tags:
      - CertCertificate
  /certCertificate/probeCertificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 域名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.ProbeCertificateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 探测成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 探测证书
      tags:
      - CertCertificate
  /certCertificate/updateAllCertificates:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量更新证书
      tags:
      - CertCertificate
  /certCertificate/updateCertCertificate:
    put:
      consumes:
      - application/json
      parameters:
      - description: 证书模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CertCertificate'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新证书
      tags:
      - CertCertificate
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...